requirements.txt
Django>=4.0
mysqlclient>=2.0.3
django-cors-headers>=3.8.0
pillow>=8.3.2
django-environ>=0.4.5
djangorestframework>=3.13
django-extensions>=3.1.3